Move data-download-related fixtures to `tests/tasks/conftest.py`
